requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13
djangorestframework-simplejwt>=5.2.0
social-auth-app-django>=5.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
mysqlclient>=2.1.0  # MySQL client for Django
python-dotenv
django-cors-headers>=3.13.0
requests
mysqlclient